scripts/requirements.txt
yt-dlp>=2024.0.0
openai-whisper>=20230314
